Reorganize button functions into separate file
