+++ src/components/stylesheet/calculator.css
@@ -27,8 +27,8 @@ body {
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  width: 20%;
-  height: 50%;
+  width: 30%;
+  height: 60%;
   background-color: var(--white);
 }
